requirements.txt
fuzzywuzzy==0.18.0
jarowinkler==1.0.2
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
